barvinok_series: remove some more equalities from the enumerated polyhedron.
remove_equalities_p does not remove equalities if they impose
some affine combination of the parameters to be a multiple of some
integer bigger than 1.
We use compress_parms to replace the current set of parameters
by a different set of parameters that have no such constraint.

After computing the generating function based on the new set
of parameters, we then substitute the original set of parameters
back in this generating function.