barvinok_count: allow polytope to be specified by its vertices.
A matrix preceded by the word "vertices" on a line by itself
will be interpreted as a list of vertices in PolyLib notation,
i.e., an n by (d+2) matrix, where n is the number of vertices/rays
and d the dimension.  The first column is 0 for lines and 1 for
vertices/rays.  The final column is the denominator or 0 for rays.
Note that for barvinok_count, the first column will always be 1.

Feature requested by Josephine Yu.