barvinok_ehrhart: compute ehrhart (quasi-)polynomial/series of a polytope.
This application provides no new functionality, but may be easier
to use for our mathematical friends.
Without the --series option, calling barvinok_ehrhart on a polytope
is equivalent to calling barvinok_enumerate on the cone generated
by the polytope.
With the --series option, calling barvinok_ehrhart on a polytope
is equivalent to calling barvinok_series on the cone generated
by the polytope.