reduced_basis: ensure all LP solutions are non-negative
glpk tends to give meaningless results when the coefficients of
the problem get too big, possibly dues to conversion errors
from arbitrary precision integers to double while setting up
the problem.
We now check that the solution of each LP problem is non-negative,
or at least not more negative than "zero" may reasonbly be expected
to be.