lexmin.cc: remove_equalities; take preimage with returned matrix
The net effect of computing this preimage should be the same as
computing the image using the other matrix.
However, this change does affect the _order_ of the constraints
in the returned polyhedron, which in turn changes the _order_
of the vertices computed by Polyhedron2Param_SimplifiedDomain.

It appears the current implementation is very sensitive to this order.
On one example, the computation time increases by 20% and the output
increases by 50% due to this change of order.