gen_fun::Hadamard_product: don't assume equalities are independent
In particular, don't try to precompute the resulting dimension,
but rather use the dimension of the first in a series of polytopes
for terms with the same denominator.

In some cases, the number of factors in the denominator may be
smaller than the dimension.  This happens when there were equalities
in the problem, but it may also happen as a result of removing
a common factor in the numerator and the denominator.
In these cases, the assumption that all equalities are independent
would result in a negative computed dimension.