volume.c: delay computation of point where volume is non-zero, if needed
Previously, we would compute this point in a given chamber
for the whole polytope.  If no such point could be found a priori that
was valid for all subsequent triangulations, we resorted to some heuristics.
Now, if no point can be found that is valid for the whole polytope,
a point is computed that is valid for a given part of the polytope.

We are now also more aggressive about skipping parts that would
have a zero volume for all integer parameter values under consideration.