reduce_domain: partition chamber complex based on inner point.
We used to subtract already seen chambers from new chambers to
ensure that no pair of chambers intersect.
However, we didn't want to create non-convex chambers, so we would
only do the subtraction when the two chambers would intersect in
a facet, but this doesn't work in general.

Instead, we now use the technique we also use in the primal decomposition
of deciding which facets are "closed", i.e., included, based on the
inner product with some inner point.
In this case, it is very important that we have the true context to
ensure that the inner point does indeed lie inside the context.