bernoulli.c: perform nested sums approximation depth first
For a piecewise polynomial, the original implementation would
first compute the common refinement of the chamber decompositions
in the sum computation of each of the pieces of the original polynomial
and then proceed to the next variable in each of the cells of this
common refinement.  This has the advantage that in each of those
cell a sum is only computed once.
However, the (current implementation of the) common refinement
computes differences of polytopes, which is rather costly.

Now, the sums are computed for each piece individually until
the deepest level and only at the very end is the common refinement
computed, in a lower-dimensional space.
This common refinement computation is therefore cheaper.
The disavantage is that a sum will be computed over the same
parameter values in different branches of the recursion tree.
However, these parameter values may be split over fewer pieces.