evalue.c: evalue_split_periods: make sure selected fractional is removed
After selecting a fractional to split, we would simply call
evalue_range_reduction_in_domain to remove the fractional
in each subdomain.
However, evalue_range_reduction_in_domain tries to be smart
and may transform the argument of the fractional before checking
if the offset is constant over the domain.  It could therefore
miss the opportunity for dropping the fract, resulting in an infinite
loop in evalue_split_periods.
(Reported by Harald Devos.)

We could simply remove the heuristic in evalue_range_reduction_in_domain,
since it is not guaranteed to work well, but this patch instead makes
sure that the selected fractional is removed by keeping track of it
explicitly.