evalue.c: evalue_sum: better handling of negative values
Commit a431c6642a113acd141f4297e9832962657b0d91 moved the splitting
into orthants into evalue_sum and made some provisions for handling
the result of summing over a negative variable.  However, the summing
itself still assumed that the variables were all non-negative.
This patch fixes this assumption by counting over v <= t <= -1
instead of over 1 <= t <= v for negative v's.

This patch also moves the conversion to floors and the shifting
of the arguments of floor into evalue_sum.
This shifting is, however, moved out of the conversion.
It may be more efficient to do the shifting while doing the conversion,
but the evalue may have been converted to floors already and
we still need to make sure the floor arguments are non-negative.
Since shifting is remove from frac2floor,
evalue_frac2floor2 and evalue_frac2floor_in_domain3 are now
redundant, but are left in because they appeared in a released version.