topcom: fix heuristic for selecting rows to use as non-negativity constraints
If one of the rows is already (the negative of) a unit vector,
we put it in the right position.  We then check the row again,
since we may not have checked the row that was on that right
position.  However, we also did this when this row had already
been checked, leading to the possibility that a pair of rows
would endlessly be swapped.
We now only reconsider the row at the current position if
the row that was moved there has not already been checked.