Polyhedron_Reduced_Basis: don't stop too early when only interested in first dir
Commit 728496985428e3db5f136acebf5118cbb934bfca
(Polyhedron_Sample: be satisfied with a reasonable choice for the first direction)
introduced two "optimizations" when using the reduced basis for
checking a set for emptiness.  First, it would stop when the
width in the first direction would become smaller than 2
and second, it would stop when we reconsider the first direction
after having considered all directions up to the last direction at least once.

This last optimization seems to be useful when the given
polytope _does_ contain integer points.  However, when
there are _no_ integer points, then at least on some examples
this "optimization" can have a disastrous effect.

We therefore remove this "optimization" until we have
a better handle on when we should apply it.