evalue_read: only accept implicit multiplication on a single line
Commit e2cad9edba2f754f1ee7dfc2f0270c6531e52f75 (evalue_read: accept
constant multiple of variable without explicit '*') allowed any
constant followed by a variable to be treated as a multiplication.
However, in a partition where the value in one part is a constant,
this constant would get combined with the constraint of the next
domain.  For example,

         v1  -9 = 0
         v2  -5 = 0
         v3  -1 = 0
         v0  -4 >= 0
         - v0 + T + 1 >= 0
         T  -10 >= 0
         - T + 50 >= 0

15
         v0  -3 = 0
         v1  -10 = 0
         v2  -5 = 0
         v3  -1 = 0
         T  -10 >= 0
         - T + 50 >= 0

7

Before, the value in the first domain would be considered
to be 15 * v0 -3.  Now, we no longer allow implicit multiplications
to cross a newline.