euler.cc: summate_over_domain: only consider actual facets
Not all constraints are active on each domain and if we don't
compute the facets of the domain, then there may be some
of these constraints left in the domain description.
Trying to find the vertices of such "facets" would obviously fail.

There is not need to remove redundant constraints from the
domain, however, as we can easily read off the facets from the vertices
active on the domain.