evalue.c: evalue_level_cmp: produce more consistent results
If two evalues are in essence the same, but not identical,
e.g., when one of them has not been reduced, then evalue_level_cmp
could in some cases return -1 on both a given call and a call
with the arguments reversed, leading to an infinite recursion
on eadd_rev.

Arguably, the arguments should be reduced before calling
evalue_level_cmp, but there is no reason it should fail this
way if they are not.