Line_Length: handle 0D polytope in 1D space
Line_Length is usually only called on full-dimensional polytopes
since equalities are removed in the very beginning.
However, it may happen that the original polytope is a product of polytopes
and that an equality among the integer points in the polytope is only
discovered after the factorization.
We could choose to check for equalities again (as we do for polytopes
living in non-1D space), but we can just as easily handle the equality
inside Line_Length.