topcom.c: compute_domain: skip vertices without domain
The number of rows and columns of a domain were taken from the
domain of the first vertex.  However, the first vertex may be
one with a lower-dimensional activity domain and the Domain fields
of such vertices are set to NULL.
Thererfore, skip vertices with NULL Domains before extracting
the number of rows and columns.