iscc: allow coercion of binary list into first element
Some operations return a binary list, where the first element is the actual
result and the second is an indication of the accurateness.
Any further manipulations are usually performed on the first element.
Although it is possible to select this first element explicitly,
it is more convenient if the element is selected implicitly, when needed.