iscc: properly handle applications of functions on sets
Before, we would just apply the function to a map with zero-dimensional
domain and the set as range.  However, this means that the result of
the application will have a zero-dimensional set as domain.
At this moment, these zero-dimensional sets are indistinguishable
from parameter sets, but this will change in the near future.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>