isl_pw_qpolynomial_sum: error out if input does not have a domain
When isl_pw_qpolynomial_sum was first created, we were unable
to detect the difference between a zero dimensional domain and
the absence of a domain.  However, we have been able to tell
the difference for a while now and it does not really make sense
to compute the sum over a non-existing domain so we no longer allow
such inputs.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>