evalue_isl.c: avoid use of isl_vec internals
We may want to hide these internals from external users at some point.
Instead of an isl_vec, we now use a Vector from PolyLib.
This should not cause any extra dependence on PolyLib as the entire
file can be removed when we stop using evalue objects.

We still assume that a PolyLib Value and an isl_int are both represented
by gmp, but this assumption was already present in the original code.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>