gen_fun::add_union: context of result should be the union of the input contexts
The original implementation would keep the context of one of the arguments
instead.

Note that this commit does not have any user visible effect since the
only call to add_union in the source is followed by a replacement
of the context.

Only users who directly call this undocumented method should be affected.

Reported-by: Matt Fredrikson <mfredrik@cs.wisc.edu>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>