reduce_evalue: reduce polynomial expressions in denominator-2 fractional parts
emul_fractionals assumes that polynomial expressions in fractional parts
of an argument with denominator 2 are always affine since they start
out being affine and emul_fractionals makes sure that they stay affine.
However, a fractional part may originally have an argument with
a denominator that is different from 2 and that only becomes two
after a substitution.  The polynomial expression in this fractional part
may already be of higher degree at this point so we need to reduce
it to maintain the invariant.

Reported-by: Thomas Kahle <tomka@gentoo.org>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>