configure.ac: allow builddir for NTL
Allow to choose between system and build in --with-ntl configure option
and to provide related path through --with-ntl-prefix or
--with-ntl-builddir.  The latter should point to the root directory of
the NTL build that contains src and include where configure was invoked
inside src (the only possibility in NTL).  The option is named builddir
to be consistent with other submodules and to suggest that NTL
should be built before building barvinok.  In case the NTL source directory
is given to this option, a suggestion to use the root directory is
emitted within the error message.  Note that NTL does not use autotools so
AX_SUBMODULE may not be used directly for it.

Signed-off-by: Oleksandr Zinenko <zinenko@lri.fr>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>