assume NTL has been compiled in ISO mode
Older versions of NTL could be compiled with either NTL_STD_CXX=on
or NTL_STD_CXX=off.  NTL_STD_CXX=on became the default in NTL 5.4
and the choice was removed completely in NTL 7.0.0.
This means that the NTL_STD_CXX define can no longer be used
to check whether NTL is using namespaces.
Instead, require that NTL has been compiled in ISO mode.

Reported-by: Thomas Kahle <thomas.kahle@jpberlin.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>