add get_submodules.sh
This script encapsulates the submodule init and submodule update and
also fetches the imath submodule of isl.

Note that it is currently not possible to use the imath backend for isl
from within barvinok since barvinok uses the GMP specific
interface of isl/val_gmp.h
However, the imath submodule is also included in the barvinok distribution,
so for consistency, it should be included in the source tree
when obtained from the git repository.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>