Polyhedron_Factor: make sure group information gets updated
Polyhedron_Factor lazily keeps track of which group a column belongs to.
The actual construction of the factors assumes the groups have been
updated, however, so make sure that the group of each column eventually
gets updated.
isl_basic_set_factorizer in isl used to have the same bug, but that one
was fixed in isl-0.05.1-6-gf49c274.

Reported-by: Diego Garbervetsky <diego.garbervetsky@gmail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>